Release v1.0: Note we now do .uf2 file releases
Add a .uf2 file release to the repository to make the project available
without having to build from source. Document that in the README.

Whilst there, also add an extra (better) picture of a built dongle in
an altoids tin, and make notes in the code and README that you can easily
swap to an external LED using GPIO9.

Signed-off-by: Graham Whaley <graham.whaley@gmail.com>